Do not depend on unversioned libsqlite3.so file
On Fedora/RHEL/CentOS, the unversioned .so files are in the -devel
packages, which should not be needed for normal usage of a package. For
now, try to find libsqlite3.so.0 first, as that seems to be the current
(and only) released versioned .so file for libsqlite.

Issue found by Alexander Bokovoy <abokovoy@redhat.com>